Moved temperature specific processes to tsm_processes file
