+++ src/clearwater_modules_python/shared/processes.py
@@ -49,239 +49,3 @@ def compute_depth(
         volume: state variable for volume of computational cell provided by CWR engine
     """
     return volume / surface_area
-
-@numba.njit
-def mf_d_esat_dT(
-    water_temp_k: float,
-    a1: float,
-    a2: float,
-    a3: float,
-    a4: float,
-    a5: float,
-    a6: float,
-) -> float:
-    """
-    Compute the derivative of function computing saturation vapor pressure 
-    as a function of water temperature (Kelvin)
-
-    Fitting parameters for vapor pressure:
-    Brutsaert (1982) Evaporation into the Atmosphere, p42
-    """
-    return (
-        a1 +
-        (2.0 * a2 * water_temp_k) +
-        (3.0 * a3 * water_temp_k**2.0) +
-        (4.0 * a4 * water_temp_k**3.0) +
-        (5.0 * a5 * water_temp_k**4.0) +
-        (6.0 * a6 * water_temp_k**5.0)
-    )
-
-
-# -----------------------------------------------------------------------------------
-# Define functions to be used in the latent heat formulation
-# -----------------------------------------------------------------------------------
-
-@numba.njit
-def mf_q_longwave_down(
-    air_temp_k: float,
-    emissivity_air: float,
-    cloudiness: float,
-    stefan_boltzmann: float,
-) -> float:
-    """
-    Compute downwelling longwave radiation (W/m2)
-
-    Parameters:
-        air_temp_k (float):              Air temperature (Kelvin)
-        emissivity_air (float):     Emissivity of air (unitless)
-        cloudiness (float):         Cloudiness (fraction)
-
-    Returns:
-        Downwelling longwave radiation (W/m2, float)
-    """
-
-    return (1.0 + 0.17 * cloudiness**2) * emissivity_air * stefan_boltzmann * air_temp_k**4.0
-
-
-@numba.njit
-def mf_q_longwave_up(
-    water_temp_k: float,
-    emissivity_water: float,
-    stefan_boltzmann: float,
-) -> float:
-    """
-    Compute upwelling longwave radiation (W/m2) as a function of water temperature (Kelvin)
-    """
-
-    # logger.debug(f'mf_q_longwave_up({water_temp_k:.2f})')
-
-    return emissivity_water * stefan_boltzmann * water_temp_k**4.0
-
-
-@numba.njit
-def mf_esat_mb(
-    water_temp_k: float,
-    a0: float,
-    a1: float,
-    a2: float,
-    a3: float,
-    a4: float,
-    a5: float,
-    a6: float,
-) -> float:
-    """
-    Compute the saturation vapor pressure as a function of water temperature (Kelvin)
-
-    Fitting parameters for vapor pressure are defined in:
-    Brutsaert (1982) Evaporation into the Atmosphere, p42.
-    """
-
-    return (
-        a0 +
-        water_temp_k *
-        (
-            a1 +
-            water_temp_k *
-            (
-                a2 +
-                water_temp_k * (
-                    a3 +
-                    water_temp_k *
-                    (a4 + water_temp_k * (a5 + water_temp_k * a6))
-                )
-            )
-        )
-    )
-
-
-# Temperature conversion functions
-
-@numba.njit
-def ri_number(
-    gravity: float,
-    density_air: float,
-    density_air_sat: float,
-    wind_speed: float,
-) -> float:
-    """Calculates the Richardson Number.
-
-    Args:
-        gravity: Gravity (m/s2)
-        density_air: Density of air (kg/m3)
-        density_air_sat: Saturation density of air (kg/m3)
-        wind_speed: Wind speed (m/s)
-    """
-    return (
-        gravity *
-        (density_air - density_air_sat) *
-        2.0 / (density_air * (wind_speed**2.0))
-    )
-
-
-@numba.njit
-def ri_function(ri_number: float) -> float:
-    """Calculates the Richardson Function from the Richardson Number.
-    Richardson Number:
-        Unstable: 0.01 >= ri_function
-        Stable: 0.01 <= ri_function < 2
-        Neutral: -0.01 <  ri_function < 0.01
-    """
-    # TODO: figure out how to make this work
-    return ri_number ** 2
-    # Set bounds
-    if ri_number > 2.0:
-        ri_number = 2.0
-    elif ri_number < -1.0:
-        ri_number = -1.0
-
-    if ri_number < 0.0:
-        if ri_number >= - 0.01:
-            # neutral
-            return 1.0
-        else:
-            # unstable
-            return (1.0 - 22.0 * ri_number)**0.80
-    else:
-        if ri_number <= 0.01:
-            # neutral
-            return 1.0
-        else:
-            # stable
-            return (1.0 + 34.0 * ri_number)**(-0.80)
-
-
-@numba.njit
-def mf_latent_heat_vaporization(water_temp_k: float) -> float:
-    """
-    Compute the latent heat of vaporization (W/m2) as a function of water temperature (Kelvin)
-    """
-
-    return 2499999 - 2385.74 * water_temp_k
-
-
-@numba.njit
-def mf_density_water(water_temp_c: float) -> float:
-    """
-    Compute density of water (kg/m3) as a function of water temperature (Celsius)
-    """
-
-    # logger.debug(f'mf_density_water({water_temp_c:.2f})')
-
-    return (
-        999.973 *
-        (1.0 - (
-            (
-                (water_temp_c - 3.9863) *
-                (water_temp_c - 3.9863) *
-                (water_temp_c + 288.9414)
-            ) /
-            (
-                508929.2 *
-                (water_temp_c + 68.12963)
-            )
-        ))
-    )
-
-
-@numba.njit
-def mf_density_air_sat(water_temp_k: float, esat_mb: float, pressure_mb: float) -> float:
-    """
-    Compute the density of saturated air at water surface temperature.
-
-    Parameters:
-        water_temp_k (float):        Water temperature (Kelvin)
-        esat_mb (float):        Saturation vapor pressure in millibars
-        pressure_mb (float):    Air pressure in millibars
-
-    Returns:
-        Density of saturated air at water surface temperature (kg/m3, float)
-    """
-
-    # logger.debug(f'mf_density_air_sat({water_temp_k:.2f}, {esat_mb:.2f}, {pressure_mb:.2f})')
-
-    mixing_ratio_sat = 0.622 * esat_mb / (pressure_mb - esat_mb)
-    return 0.348 * (pressure_mb / water_temp_k) * (1.0 + mixing_ratio_sat) / (1.0 + 1.61 * mixing_ratio_sat)
-
-
-@numba.njit
-def mf_cp_water(water_temp_c: float) -> float:
-    """
-    Compute the specific heat of water (J/kg/K) as a function of water temperature (Celsius).
-    This is used in computing the source/sink term.
-    """
-    # TODO: make this work as a ufunc
-    return water_temp_c * 2
-    if water_temp_c <= 0.0:
-        return 4218.0
-    elif water_temp_c <= 5.0:
-        return 4202.0
-    elif water_temp_c <= 10.0:
-        return 4192.0
-    elif water_temp_c <= 15.0:
-        return 4186.0
-    elif water_temp_c <= 20.0:
-        return 4182.0
-    elif water_temp_c <= 25.0:
-        return 4180.0
-    else:
-        return 4178.0
