+++ src/clearwater_modules_python/tsm/dynamic_variables.py
@@ -1,7 +1,7 @@
 import clearwater_modules_python.shared.processes as shared_processes
 from clearwater_modules_python import base
 from clearwater_modules_python.tsm.model import EnergyBudget
-from clearwater_modules_python.tsm import processes
+from clearwater_modules_python.tsm import tsm_processes
 
 
 @base.register_variable(models=EnergyBudget)
@@ -15,7 +15,7 @@ Variable(
     units='K',
     description='Air temperature',
     use='dynamic',
-    process=processes.air_temp_k,
+    process=tsm_processes.air_temp_k,
 )
 Variable(
     name='water_temp_k',
@@ -23,7 +23,7 @@ Variable(
     units='K',
     description='Water temperature',
     use='dynamic',
-    process=processes.water_temp_k,
+    process=tsm_processes.water_temp_k,
 )
 Variable(
     name='mixing_ratio_air',
@@ -31,7 +31,7 @@ Variable(
     units='unitless',
     description='Mixing ratio of air',
     use='dynamic',
-    process=processes.mixing_ratio_air,
+    process=tsm_processes.mixing_ratio_air,
 )
 Variable(
     name='density_air',
@@ -39,7 +39,7 @@ Variable(
     units='kg/m^3',
     description='Density of air',
     use='dynamic',
-    process=processes.density_air,
+    process=tsm_processes.density_air,
 )
 Variable(
     name='density_water',
@@ -47,7 +47,7 @@ Variable(
     units='kg/m^3',
     description='Density of water',
     use='dynamic',
-    process=shared_processes.mf_density_water,
+    process=tsm_processes.mf_density_water,
 )
 Variable(
     name='esat_mb',
@@ -55,7 +55,7 @@ Variable(
     units='mb',
     description='Saturation vapor pressure',
     use='dynamic',
-    process=shared_processes.mf_esat_mb,
+    process=tsm_processes.mf_esat_mb,
 )
 Variable(
     name='density_air_sat',
@@ -63,7 +63,7 @@ Variable(
     units='kg/m^3',
     description='Density of air at saturation',
     use='dynamic',
-    process=shared_processes.mf_density_air_sat,
+    process=tsm_processes.mf_density_air_sat,
 )
 Variable(
     name='ri_number',
@@ -71,7 +71,7 @@ Variable(
     units='unitless',
     description='Richardson number',
     use='dynamic',
-    process=shared_processes.ri_number,
+    process=tsm_processes.ri_number,
 )
 Variable(
     name='ri_function',
@@ -79,7 +79,7 @@ Variable(
     units='unitless',
     description='Richardson function',
     use='dynamic',
-    process=shared_processes.ri_function,
+    process=tsm_processes.ri_function,
 )
 Variable(
     name='lv',
@@ -87,7 +87,7 @@ Variable(
     units='J/kg',
     description='Latent heat of vaporization',
     use='dynamic',
-    process=shared_processes.mf_latent_heat_vaporization,
+    process=tsm_processes.mf_latent_heat_vaporization,
 )
 Variable(
     name='cp_water',
@@ -95,7 +95,7 @@ Variable(
     units='J/kg/K',
     description='Specific heat of water',
     use='dynamic',
-    process=shared_processes.mf_cp_water,
+    process=tsm_processes.mf_cp_water,
 )
 Variable(
     name='emissivity_air',
@@ -103,7 +103,7 @@ Variable(
     units='unitless',
     description='Emissivity of air',
     use='dynamic',
-    process=processes.emissivity_air,
+    process=tsm_processes.emissivity_air,
 )
 Variable(
     name='wind_function',
@@ -111,7 +111,7 @@ Variable(
     units='unitless',
     description='Wind function',
     use='dynamic',
-    process=processes.wind_function,
+    process=tsm_processes.wind_function,
 )
 Variable(
     name='q_latent',
@@ -119,7 +119,7 @@ Variable(
     units='W/m^2',
     description='Latent heat flux',
     use='dynamic',
-    process=processes.q_latent,
+    process=tsm_processes.q_latent,
 )
 Variable(
     name='q_sensible',
@@ -127,7 +127,7 @@ Variable(
     units='W/m^2',
     description='Sensible heat flux',
     use='dynamic',
-    process=processes.q_sensible,
+    process=tsm_processes.q_sensible,
 )
 Variable(
     name='q_sediment',
@@ -135,7 +135,7 @@ Variable(
     units='W/m^2',
     description='Sediment heat flux',
     use='dynamic',
-    process=processes.q_sediment,
+    process=tsm_processes.q_sediment,
 )
 Variable(
     name='dTdt_sediment_c',
@@ -143,7 +143,7 @@ Variable(
     units='degC',
     description='Sediment temperature change',
     use='dynamic',
-    process=processes.dTdt_sediment_c,
+    process=tsm_processes.dTdt_sediment_c,
 )
 Variable(
     name='q_net',
@@ -151,7 +151,7 @@ Variable(
     units='W/m^2',
     description='Net heat flux',
     use='dynamic',
-    process=processes.q_net,
+    process=tsm_processes.q_net,
 )
 Variable(
     name='q_longwave_down',
@@ -159,7 +159,7 @@ Variable(
     units='W/m2',
     description='Downwelling longwave radiation',
     use='dynamic',
-    process=shared_processes.mf_q_longwave_down,
+    process=tsm_processes.mf_q_longwave_down,
 )
 
 Variable(
@@ -168,7 +168,7 @@ Variable(
     units='W/m2',
     description='Upwelling longwave radiation',
     use='dynamic',
-    process=shared_processes.mf_q_longwave_up,
+    process=tsm_processes.mf_q_longwave_up,
 )
 
 Variable(
@@ -177,5 +177,5 @@ Variable(
     units='degC',
     description='Water temperature change',
     use='dynamic',
-    process=processes.dTdt_water_c,
+    process=tsm_processes.dTdt_water_c,
 )
